Lab3
I'm really upset that I cannot find a way to separate animations for the jump.  At minimum there are 4 I would like to trigger, the takeoff, the idle fly, the flap on "Jump" buttonPress, and the land.  It doesn't look bad transitioning without those, but it doesn't look great either.  And I still haven't figured out how to tie the flap animation to the buttonPress, I tried, it's still in the PlayerController C# file, I'm just frustrated and ready to move on to the firing animations and functions.